refactor: dedup code
